Move of some components in the right folder
