+++ app/src/androidTest/java/ch/epfl/skysync/components/TimerTest.kt
@@ -7,7 +7,6 @@ import androidx.compose.ui.test.hasText
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
-import ch.epfl.skysync.ui.components.Timer
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
