+++ app/src/androidTest/java/ch/epfl/skysync/components/flightFormTest.kt
@@ -15,6 +15,7 @@ import androidx.compose.ui.test.performScrollToNode
 import androidx.compose.ui.test.performTextClearance
 import androidx.compose.ui.test.performTextInput
 import androidx.navigation.testing.TestNavHostController
+import ch.epfl.skysync.components.forms.FlightForm
 import ch.epfl.skysync.models.flight.Balloon
 import ch.epfl.skysync.models.flight.BalloonQualification
 import ch.epfl.skysync.models.flight.Basket
