+++ app/src/main/java/ch/epfl/skysync/components/Timer.kt
@@ -1,4 +1,4 @@
-package ch.epfl.skysync.ui.components
+package ch.epfl.skysync.components
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
