+++ app/src/main/java/ch/epfl/skysync/screens/AddFlight.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavHostController
-import ch.epfl.skysync.components.FlightForm
+import ch.epfl.skysync.components.forms.FlightForm
 import ch.epfl.skysync.models.flight.PlannedFlight
 import ch.epfl.skysync.models.flight.RoleType
 import ch.epfl.skysync.viewmodel.FlightsViewModel
